+++ packages/core-json-rpc/lib/server/methods/wallets/bip38/create.js
@@ -4,6 +4,7 @@ const bip39 = require('bip39')
 const bip38 = require('bip38')
 const database = require('../../../services/database')
 const getBIP38Wallet = require('../../../utils/bip38-keys')
+const decryptWIF = require('../../../utils/decrypt-wif')
 
 module.exports = {
   name: 'wallets.bip38.create',
@@ -19,20 +20,22 @@ module.exports = {
     } catch (error) {
       const { publicKey, privateKey } = crypto.getKeys(bip39.generateMnemonic())
 
-      const encryptedWif = bip38.encrypt(
+      const encryptedWIF = bip38.encrypt(
         Buffer.from(privateKey, 'hex'),
         true,
         params.bip38 + params.userId,
       )
       await database.set(
         utils.sha256(Buffer.from(params.userId)).toString('hex'),
-        encryptedWif,
+        encryptedWIF,
       )
+
+      const { wif } = decryptWIF(encryptedWIF, params.userId, params.bip38)
 
       return {
         publicKey,
         address: crypto.getAddress(publicKey),
-        wif: encryptedWif,
+        wif,
       }
     }
   },